+++ R/package_tools.R
@@ -2619,10 +2619,10 @@ sn_install_shennong <- function(
 
 .sn_bioconductor_packages <- function() {
   c(
-    "Banksy", "BiocParallel", "celda", "clusterProfiler", "Coralysis", "decoupleR", "DESeq2", "dorothea", "edgeR", "GENIE3",
+    "apeglm", "Banksy", "BiocParallel", "celda", "clusterProfiler", "Coralysis", "decoupleR", "DESeq2", "dorothea", "edgeR", "GENIE3",
     "glmGamPoi", "miloR", "nnSVG", "org.Hs.eg.db", "org.Mm.eg.db", "progeny", "rhdf5",
     "rtracklayer", "S4Vectors", "scDblFinder", "scDesign3", "scran", "SingleCellExperiment", "SpatialExperiment", "Nebulosa",
-    "SummarizedExperiment", "limma"
+    "SummarizedExperiment", "limma", "variancePartition"
   )
 }